some refactoring and simplifying #862
