+++ app/helpers/resource_list_item_helper.rb
@@ -266,77 +266,6 @@ module ResourceListItemHelper
     end
   end
 
-  def list_view_table_row(resource, tableview_columns)
-    content_tag :tr, class: :list_item do
-      tableview_columns.collect do |column|
-        content_tag :td do
-          list_view_table_column(resource, column)
-        end
-      end.join.html_safe
-    end
-  end
-
-  def list_view_table_column(resource, column)
-    case column
-    when 'title'
-      list_item_title resource, {}
-    when 'projects'
-      if resource.respond_to?('projects')
-        resource.projects.uniq.map { |p| link_to p.title, p }.join(', ').html_safe
-      else
-        'Field not found'
-      end
-    when 'contributor'
-      if resource.respond_to?('contributor')
-        link_to resource.contributor.name, resource.contributor
-      else
-        'Field not found'
-      end
-    when 'tags'
-      if resource.respond_to?('tags')
-        resource.send(column).join(',')
-      else
-        'Field not found'
-      end
-    when 'creators'
-      table_item_person_list resource.creators
-    when 'assay_type_uri'
-      link_to_assay_type(resource)
-    when 'technology_type_uri'
-      link_to_technology_type(resource)
-    else
-      if resource.respond_to?(column)
-        text_or_not_specified(resource.send(column), length: 300, auto_link: true, none_text: '')
-      else
-        'Unknown Column parameter'
-      end
-    end
-  end
-
-  def table_item_person_list(contributors, other_contributors = nil, key = t('creator').capitalize)
-    contributor_count = contributors.count
-    contributor_count += 1 unless other_contributors.blank?
-    html = ''
-    other_html = ''
-    if (key == 'Author')
-      html << contributors.map do |author|
-        if author.person
-          link_to author.full_name, show_resource_path(author.person)
-        else
-          author.full_name
-        end
-      end.join(', ')
-    else
-      html << contributors.map {|c| link_to truncate(c.title, length: 75), show_resource_path(c), title: get_object_title(c)}.join(', ')
-    end
-    unless other_contributors.blank?
-      other_html << ', ' unless contributors.empty?
-      other_html << other_contributors
-    end
-    other_html << 'None' if contributor_count == 0
-    html.html_safe + other_html
-  end
-
   def list_item_author_list(all_authors)
     list_item_person_list(all_authors, nil, 'Author')
   end
