+++ app/views/assets/_resource_list_item_table.html.erb
@@ -20,5 +20,5 @@
 %>
 
 
-<%= list_view_table_row(resource, tableview_columns) %>
+<%= resource_list_table_row(resource, tableview_columns) %>
